requirements.txt
python-telegram-bot==21.0.1
py-tgcalls==2.2.5
pyrogram==2.0.106
TgCrypto==1.2.5
yt-dlp==2023.12.30
youtube-search-python==1.6.6
aiofiles==23.2.1
aiohttp==3.9.1
asyncio==3.4.3
psutil==5.9.6
motor==3.3.2
dnspython==2.4.2
lyricsgenius==3.0.1
spotipy==2.22.1
requests==2.31.0
Pillow==10.2.0
mutagen==1.47.0
ffmpeg-python==0.2.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-dotenv==1.0.0
colorlog==6.8.0
psycopg2-binary==2.9.9